added uniqueness check

+++ src/main/resources/db/changelog/v.1.0/sql/create-amenity_hotel-table.sql
@@ -1,7 +1,7 @@
 CREATE TABLE amenity_hotel
 (
-    hotel_id   INT,
-    amenity_id INT,
+    hotel_id   BIGINT,
+    amenity_id BIGINT,
     PRIMARY KEY (hotel_id, amenity_id),
     FOREIGN KEY (hotel_id) REFERENCES hotel (id),
     FOREIGN KEY (amenity_id) REFERENCES amenity (id)
